Add gem requirement to package description
Until we finish converting the core to JS, we will need people to
install the gem. Including this information in the description will show
people this requirement at the time they install the plugin in their
editors, which I think will help smooth things out for people getting
set up for the first time.