Fix skipping matched imports when asking for single selection
If I have a file that has two imports that need to be added, one with
multiple possibilities and the other with a single match, after
selecting the import from the list only the selected import is added
while the one with a single match is skipped. If there are multiple
imports to be selected, things work as expected.

This is because we are passing the --word flag if only one selection
needed to be made instead of when we are importing a word. This commit
fixes this by carrying the word through the execution chain.

Fixes #1