Begin conversion to JS version of import-js
We have converted import-js to be in JavaScript instead of Ruby. This
means that we can depend on in directly instead of shelling out. At the
same time, we've made some adjustments to the API, which makes things a
little more straightforward and performant.

I have started the conversion in this commit. Importing a word seems to
be working somewhat okay. I still need to make fix imports work
correctly, and I haven't tried goto. There may also be some subtle bugs
that will need to be shaken out.