Fix creating success notifications from result messages
This is now an array, so we need to either convert it to a string or add
multiple messages. For now, I went with the easy approach of just
converting it to a string. If we find out that it would be better to do
something else, we can go down that path in then.