Simplify asking for selections
Now that we pass the paths to import-js instead of the selection
indexes, this can be done very quickly so we may as well simplify things
a bit and update the buffer after resolving each import. This will make
sure that the editor is always as up to date as it can be and prevent it
from losing changes if something catastrophic happens midway through.