Upgrade import-js to 1.0.0
The new version of import-js has support for parallel file finding, and
because of that all import related functions return promises instead of
values.

While making the code changes necessary for the upgrade, I think I found
an issue with how we change `cwd` in between import calls. Import-js
isn't currently designed for switching cwd midway. I'll look into that
so that we can finish the upgrade. Switching cwd will also be a
prerequisite for enabling the watchman file cache.