Use Watchman if available
Import-js now integrates with Watchman [1]. Since we don't use the
`importjsd` command in the Atom plugin (which other editors do), we need
to manually initialize the watchman file cache. I decided to lazily
initialize the watchman cache on each working directory. This means that
the first import will be a little slower than the rest (because we
haven't built up the cache yet).

The watchman integration is smart enough to fallback to a non-watchman
file finder if watchman isn't installed. We'll only attempt to
initialize once. If initialization fails (due to e.g. watchman not being
installed) we won't try again.

An unrelated change snuck in:
Adding an .importjs.json file so that importing works correctly (we
can't use es2015 import statements).

[1]: https://facebook.github.io/watchman/