Update unresolved imports flow for import-js 2.7.0
A recent commit in import-js made it so that we send down more
information about unresolved imports:
https://github.com/Galooshi/import-js/commit/194ed526a2ab0b491c54afdd920e08f7f4332424

This was in service of allowing a module to export named export equal to
the module itself. Before said commit, we didn't have enough information
in the json object sent down to the client to target the right import.
Now, we have a full object describing the import.